utils/password.go
package utils

import (
	"fmt"
	"golang.org/x/crypto/bcrypt"
)

// HashPassword returns the password hashed by bcrypt
func HashPassword(password string) (string, error) {
	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
	if err != nil {
		return "", fmt.Errorf("fail to hash password: %w", err)
	}

	return string(hashedPassword), nil
}

// CheckPassword check if the password is match or not
func CheckPassword(password, hashedPassword string) error {
	return bcrypt.CompareHashAndPassword([]byte(hashedPassword), []byte(password))
}
